[algo] feat: support router replay (#4101)
### What does this PR do?

This PR introduces a draft **Router Replay** support into Verl.
Inspired by the recent research in **MoE Reinforcement
Learning**([2510.11370](https://arxiv.org/abs/2510.11370),
[2507.18071](https://arxiv.org/abs/2507.18071)), this implementation
supports **Router Replay (R2)** and **Rollout Router Replay (R3)**.
R2 allows recording routing token selection during` log probability
computation` and replaying expert selection during policy update. R3
enables recording during `model inference` and replaying during RL
post-training.

The initial version supports **Router Replay** with `Megatron` backend,
including comprehensive support for distributed training strategies
(**DP, TP, EP, ETP, PP, and Re-compute**).


The current implementation uses a patch-based approach. Once the
upstream PR
[NVIDIA/Megatron-LM#2101](https://github.com/NVIDIA/Megatron-LM/pull/2101)
is merged or provides corresponding interfaces, the patch can be removed
and replaced with official API integration.

## Usage Tutorial

### Basic Configuration
To enable Router Replay functionality, add the following configuration
to your trainer config:
#### Method 1: Trainer Configuration
Add the following configuration to your trainer config:

```yaml
router_replay:
  enabled: true
  mode: "R2"  # Options: "R2", "R3"
```

#### Method 2: Launch Script Configuration
Add the following parameter to your launch script:

```bash
# In your launch script
actor_rollout_ref.actor.router_replay.mode="R2"
```

### R2 Mode Usage
1. **Enable R2 mode** in configuration
2. **Record phase**: During log probability computation, routing
selections are automatically recorded
3. **Replay phase**: During policy update, recorded expert selections
are replayed

### R3 Mode Usage
1. **Enable R3 mode** in configuration
2. **Record phase**: During model inference, routing decisions are
captured
3. **Replay phase**: During RL post-training, recorded routing data is
used
4. 
## In Progress
R2
- [ ]  FSDP backend

R3
- [x] vLLM Rollout
- [ ] Sglang Rollout

---------

+++ verl/models/mcore/util.py
@@ -144,7 +144,7 @@ def postprocess_packed_seqs(
     if cp_size > 1:
         # output shape: [1, packed_len, hidden_dim]
         # need to gather across cp group and concatenate in sequence dimension
-        output_list = [torch.empty_like(output) for _ in range(cp_size)]
+        output_list = [torch.empty_like(output, dtype=output.dtype) for _ in range(cp_size)]
         torch.distributed.all_gather(output_list, output.detach(), group=mpu.get_context_parallel_group())
         output_list[mpu.get_context_parallel_rank()] = output
     else:
@@ -159,7 +159,7 @@ def postprocess_packed_seqs(
         half_seqlen = s_len_padded_chunk // 2
         s_len = seq_lens_cpu[i]
         s_len_padded = s_len_padded_chunk * cp_size
-        tmp = torch.empty(s_len_padded, *output.shape[2:], device=output.device)
+        tmp = torch.empty(s_len_padded, *output.shape[2:], device=output.device, dtype=output.dtype)
         for j in range(cp_size):
             o = output_list[j][0]
             # split to 2 chunks
